src/requirements.txt
Flask==1.1.2
Pillow==7.1.1